Create race.c
A competition involving a group of x hikers (where x ranges from 1 to 100) revolves around carrying a flag a distance of 10 steps. Each individual hiker is instantiated as a new thread. Should a hiker drop their flag, it is denoted by the symbol "*". Upon realizing they dropped the flag, the hiker must backtrack to the point of flag drop and then resume the race from that spot.